Turned recurrent into a package to better organize models. Public interface of recurrent is unchanged.
