+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "deepsysid"
-version = "0.1.12"
+version = "0.1.13"
 description = "A system identification toolkit for multistep prediction using deep learning and hybrid methods."
 authors = ["Alexandra Baier <alx.baier@gmail.com>", "Daniel Frank <daniel@familiefrank.eu>"]
 readme = "README.md"
