+++ src/deepsysid/models/recurrent/__init__.py
@@ -1,0 +1,31 @@
+from .constrained_rnn import (
+    ConstrainedRnn,
+    ConstrainedRnnConfig,
+    HybridConstrainedRnn,
+    HybridConstrainedRnnConfig,
+    LtiRnnInit,
+    LtiRnnInitConfig,
+    RnnInitFlexibleNonlinearity,
+    RnnInitFlexibleNonlinearityConfig,
+)
+from .separate_initialization import (
+    LSTMInitModel,
+    LSTMInitModelConfig,
+    RnnInit,
+    RnnInitConfig,
+)
+
+__all__ = [
+    'RnnInitFlexibleNonlinearityConfig',
+    'RnnInitFlexibleNonlinearity',
+    'LtiRnnInitConfig',
+    'LtiRnnInit',
+    'ConstrainedRnnConfig',
+    'ConstrainedRnn',
+    'HybridConstrainedRnnConfig',
+    'HybridConstrainedRnn',
+    'RnnInitConfig',
+    'RnnInit',
+    'LSTMInitModelConfig',
+    'LSTMInitModel',
+]
